Fix xcode 11 build issues.

+++ MSAL/src/MSALPublicClientApplication.m
@@ -467,7 +467,7 @@
                                                      oidcScopes:[self.class defaultOIDCScopes]
                                            extraScopesToConsent:parameters.extraScopesToConsent ? [[NSOrderedSet alloc]     initWithArray:parameters.extraScopesToConsent copyItems:YES] : nil
                                                   correlationId:parameters.correlationId
-                                                 telemetryApiId:[NSString stringWithFormat:@"%lu", parameters.telemetryApiId]
+                                                 telemetryApiId:[NSString stringWithFormat:@"%ld", (long)parameters.telemetryApiId]
                                         supportedBrokerProtocol:MSID_BROKER_MSAL_SCHEME
                                                     requestType:interactiveRequestType
                                                           error:&msidError];
@@ -844,7 +844,7 @@
                                                                               scopes:[[NSOrderedSet alloc] initWithArray:parameters.scopes copyItems:YES]
                                                                           oidcScopes:[self.class defaultOIDCScopes]
                                                                        correlationId:parameters.correlationId
-                                                                      telemetryApiId:[NSString stringWithFormat:@"%lu", parameters.telemetryApiId]
+                                                                          telemetryApiId:[NSString stringWithFormat:@"%ld", (long)parameters.telemetryApiId]
                                                                                error:&msidError];
     
     if (!msidParams)
